+++ Source/WebCore/Configurations/WebCore.xcconfig
@@ -241,10 +241,33 @@ SWIFT_OBJC_INTEROP_MODE = objcxx;
 MODULEMAP_FILE = $(SRCROOT)/WebCore.modulemap;
 MODULEMAP_PRIVATE_FILE = $(SRCROOT)/WebCore_Private.modulemap;
 
-// Temporarily disabled until rdar://166630726 is fixed, which is downstream of
-// PAL modularization.
-ENABLE_MODULE_VERIFIER = NO;
-ENABLE_MODULE_VERIFIER[sdk=macosx14*] = NO;
+WK_VERIFY_WEBCORE_MODULE = YES;
+
+WK_VERIFY_WEBCORE_MODULE[sdk=macosx14*] = ;
+WK_VERIFY_WEBCORE_MODULE[sdk=macosx15*] = ;
+WK_VERIFY_WEBCORE_MODULE[sdk=macosx26*] = ;
+
+WK_VERIFY_WEBCORE_MODULE[sdk=iphone*26.0*] = ;
+WK_VERIFY_WEBCORE_MODULE[sdk=iphone*26.1*] = ;
+WK_VERIFY_WEBCORE_MODULE[sdk=iphone*26.2*] = ;
+WK_VERIFY_WEBCORE_MODULE[sdk=iphone*26.3*] = ;
+
+WK_VERIFY_WEBCORE_MODULE[sdk=appletv*26.0*] = ;
+WK_VERIFY_WEBCORE_MODULE[sdk=appletv*26.1*] = ;
+WK_VERIFY_WEBCORE_MODULE[sdk=appletv*26.2*] = ;
+WK_VERIFY_WEBCORE_MODULE[sdk=appletv*26.3*] = ;
+
+WK_VERIFY_WEBCORE_MODULE[sdk=watch*26.0*] = ;
+WK_VERIFY_WEBCORE_MODULE[sdk=watch*26.1*] = ;
+WK_VERIFY_WEBCORE_MODULE[sdk=watch*26.2*] = ;
+WK_VERIFY_WEBCORE_MODULE[sdk=watch*26.3*] = ;
+
+WK_VERIFY_WEBCORE_MODULE[sdk=xr*26.0*] = ;
+WK_VERIFY_WEBCORE_MODULE[sdk=xr*26.1*] = ;
+WK_VERIFY_WEBCORE_MODULE[sdk=xr*26.2*] = ;
+WK_VERIFY_WEBCORE_MODULE[sdk=xr*26.3*] = ;
+
+ENABLE_MODULE_VERIFIER = $(WK_VERIFY_WEBCORE_MODULE);
 MODULE_VERIFIER_SUPPORTED_LANGUAGES = objective-c++;
 MODULE_VERIFIER_SUPPORTED_LANGUAGES[sdk=macosx14*] = ;
 MODULE_VERIFIER_SUPPORTED_LANGUAGE_STANDARDS = c++23;
@@ -257,7 +280,7 @@ SWIFT_DEFAULT_ACTOR_ISOLATION = nonisolated;
 SWIFT_STRICT_CONCURRENCY = complete;
 DEFINES_MODULE = YES;
 
-OTHER_MODULE_VERIFIER_FLAGS = -- -iframework $(BUILT_PRODUCTS_DIR)/SDKAdditions/System/Library/Frameworks -I$(BUILT_PRODUCTS_DIR)$(WK_LIBRARY_HEADERS_FOLDER_PATH) -I$(BUILT_PRODUCTS_DIR)$(WK_LIBRARY_HEADERS_FOLDER_PATH)/webrtc -I$(WK_PRIVATE_SDK_DIR)$(WK_ALTERNATE_WEBKIT_SDK_PATH)$(WK_LIBRARY_HEADERS_FOLDER_PATH)/webrtc  -I$(WK_ADDITIONAL_SDKS)$(WK_LIBRARY_HEADERS_FOLDER_PATH) -Xclang -fno-implicit-module-maps $(OTHER_MODULE_VERIFIER_FLAGS_$(WK_EMPTY_$(WK_OVERRIDE_FRAMEWORKS_DIR)));
+OTHER_MODULE_VERIFIER_FLAGS = -- -iframework $(BUILT_PRODUCTS_DIR)/SDKAdditions/System/Library/Frameworks -I$(BUILT_PRODUCTS_DIR)$(WK_LIBRARY_HEADERS_FOLDER_PATH) -I$(BUILT_PRODUCTS_DIR)$(WK_LIBRARY_HEADERS_FOLDER_PATH)/webrtc -I$(WK_PRIVATE_SDK_DIR)$(WK_ALTERNATE_WEBKIT_SDK_PATH)$(WK_LIBRARY_HEADERS_FOLDER_PATH)/webrtc  -I$(WK_ADDITIONAL_SDKS)$(WK_LIBRARY_HEADERS_FOLDER_PATH) $(OTHER_MODULE_VERIFIER_FLAGS_$(WK_EMPTY_$(WK_OVERRIDE_FRAMEWORKS_DIR)));
 OTHER_MODULE_VERIFIER_FLAGS_ = -iframework "$(WK_PRIVATE_SDK_DIR)$(WK_OVERRIDE_FRAMEWORKS_DIR)" -isystem "$(WK_PRIVATE_SDK_DIR)$(WK_ALTERNATE_WEBKIT_SDK_PATH)$(WK_LIBRARY_HEADERS_FOLDER_PATH)";
 
 HEADERMAP_INCLUDES_FRAMEWORK_ENTRIES_FOR_TARGETS_NOT_BEING_BUILT = NO;
