+++ Source/WebCore/WebCore_Private.modulemap
@@ -1,8 +1,4 @@
 framework module WebCore_Private [system] {
-    explicit module WebCoreVirtualLogFunctions {
-        textual header "WebCoreVirtualLogFunctions.h"
-    }
-
     explicit module PlatformExportMacros {
         header "PlatformExportMacros.h"
         export *
