+++ Source/WebCore/platform/graphics/cocoa/MediaPlayerPrivateWebM.h
@@ -32,9 +32,9 @@
 #include <WebCore/HTMLMediaElementIdentifier.h>
 #include <WebCore/MediaPlayerPrivate.h>
 #include <WebCore/PlatformLayer.h>
-#include <WebCore/SourceBufferParserWebM.h>
 #include <WebCore/TimeRanges.h>
 #include <WebCore/VideoFrameMetadata.h>
+#include "SourceBufferParserWebM.h"
 #include "WebMResourceClient.h"
 #include <wtf/HashFunctions.h>
 #include <wtf/LoggerHelper.h>
