+++ Source/WebCore/platform/mediastream/libwebrtc/LibWebRTCRefWrappers.h
@@ -28,6 +28,9 @@
 
 #include <WebCore/LibWebRTCMacros.h>
 
+// FIXME: Modularize WebRTC
+IGNORE_CLANG_WARNINGS_BEGIN("non-modular-include-in-module")
+
 WTF_IGNORE_WARNINGS_IN_THIRD_PARTY_CODE_BEGIN
 IGNORE_CLANG_WARNINGS_BEGIN("nullability-completeness")
 #include <webrtc/api/data_channel_interface.h>
@@ -43,6 +46,8 @@ IGNORE_CLANG_WARNINGS_BEGIN("nullability-completeness")
 #include <webrtc/api/sctp_transport_interface.h>
 IGNORE_CLANG_WARNINGS_END
 WTF_IGNORE_WARNINGS_IN_THIRD_PARTY_CODE_END
+
+IGNORE_CLANG_WARNINGS_END
 
 #include <wtf/Ref.h>
 #include <wtf/RefPtr.h>
