+++ Source/WebCore/platform/mediastream/libwebrtc/VideoFrameLibWebRTC.h
@@ -33,8 +33,11 @@
 
 WTF_IGNORE_WARNINGS_IN_THIRD_PARTY_CODE_BEGIN
 
+// FIXME: Modularize WebRTC
+IGNORE_CLANG_WARNINGS_BEGIN("non-modular-include-in-module")
 #include <webrtc/api/video/video_frame.h>
 #include <webrtc/webkit_sdk/WebKit/WebKitUtilities.h>
+IGNORE_CLANG_WARNINGS_END
 
 WTF_IGNORE_WARNINGS_IN_THIRD_PARTY_CODE_END
